requirements.txt
# =============================================================================
# Text & Table Builder v1.2 - Python Dependencies
# =============================================================================

# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.0

# Google Cloud / Vertex AI Integration
google-cloud-aiplatform>=1.70.0
google-auth>=2.35.0
vertexai>=1.70.0

# Async Support
aiohttp>=3.10.0
httpx>=0.27.0

# OpenAI (optional - for future multi-provider support)
openai>=1.50.0

# Anthropic (optional - for future multi-provider support)
anthropic>=0.39.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0

# Development and Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Type Hints
types-requests>=2.32.0

# =============================================================================
# Installation Instructions
# =============================================================================
#
# Standard installation:
#   pip install -r requirements.txt
#
# Development installation (includes testing):
#   pip install -r requirements.txt
#
# For production, pin versions exactly as shown above.
# For development, you can use >= for more flexibility.
#
# =============================================================================
